Added a simple import test that should be passed before anything is committed to the repository. Renamed the old import tests to DesignVerificationTests to be more descriptive of what they are doing.
